+++ lib/private/legacy/user/backend.php
@@ -62,137 +62,6 @@ define('OC_USER_BACKEND_COUNT_USERS',       1 << 28);
  * Abstract base class for user management. Provides methods for querying backend
  * capabilities.
  */
-abstract class OC_User_Backend implements \OCP\UserInterface {
-	/**
-	 * error code for functions not provided by the user backend
-	 */
-	const NOT_IMPLEMENTED = -501;
+abstract class OC_User_Backend extends \OC\User\Backend implements \OCP\UserInterface {
 
-	/**
-	 * actions that user backends can define
-	 */
-	const CREATE_USER		= 1;			// 1 << 0
-	const SET_PASSWORD		= 16;			// 1 << 4
-	const CHECK_PASSWORD	= 256;			// 1 << 8
-	const GET_HOME			= 4096;			// 1 << 12
-	const GET_DISPLAYNAME	= 65536;		// 1 << 16
-	const SET_DISPLAYNAME	= 1048576;		// 1 << 20
-	const PROVIDE_AVATAR	= 16777216;		// 1 << 24
-	const COUNT_USERS		= 268435456;	// 1 << 28
-
-	protected $possibleActions = array(
-		self::CREATE_USER => 'createUser',
-		self::SET_PASSWORD => 'setPassword',
-		self::CHECK_PASSWORD => 'checkPassword',
-		self::GET_HOME => 'getHome',
-		self::GET_DISPLAYNAME => 'getDisplayName',
-		self::SET_DISPLAYNAME => 'setDisplayName',
-		self::PROVIDE_AVATAR => 'canChangeAvatar',
-		self::COUNT_USERS => 'countUsers',
-	);
-
-	/**
-	* Get all supported actions
-	* @return int bitwise-or'ed actions
-	*
-	* Returns the supported actions as int to be
-	* compared with self::CREATE_USER etc.
-	*/
-	public function getSupportedActions() {
-		$actions = 0;
-		foreach($this->possibleActions AS $action => $methodName) {
-			if(method_exists($this, $methodName)) {
-				$actions |= $action;
-			}
-		}
-
-		return $actions;
-	}
-
-	/**
-	* Check if backend implements actions
-	* @param int $actions bitwise-or'ed actions
-	* @return boolean
-	*
-	* Returns the supported actions as int to be
-	* compared with self::CREATE_USER etc.
-	*/
-	public function implementsActions($actions) {
-		return (bool)($this->getSupportedActions() & $actions);
-	}
-
-	/**
-	 * delete a user
-	 * @param string $uid The username of the user to delete
-	 * @return bool
-	 *
-	 * Deletes a user
-	 */
-	public function deleteUser( $uid ) {
-		return false;
-	}
-
-	/**
-	 * Get a list of all users
-	 *
-	 * @param string $search
-	 * @param null|int $limit
-	 * @param null|int $offset
-	 * @return string[] an array of all uids
-	 */
-	public function getUsers($search = '', $limit = null, $offset = null) {
-		return array();
-	}
-
-	/**
-	* check if a user exists
-	* @param string $uid the username
-	* @return boolean
-	*/
-	public function userExists($uid) {
-		return false;
-	}
-
-	/**
-	* get the user's home directory
-	* @param string $uid the username
-	* @return boolean
-	*/
-	public function getHome($uid) {
-		return false;
-	}
-
-	/**
-	 * get display name of the user
-	 * @param string $uid user ID of the user
-	 * @return string display name
-	 */
-	public function getDisplayName($uid) {
-		return $uid;
-	}
-
-	/**
-	 * Get a list of all display names and user ids.
-	 *
-	 * @param string $search
-	 * @param string|null $limit
-	 * @param string|null $offset
-	 * @return array an array of all displayNames (value) and the corresponding uids (key)
-	 */
-	public function getDisplayNames($search = '', $limit = null, $offset = null) {
-		$displayNames = array();
-		$users = $this->getUsers($search, $limit, $offset);
-		foreach ( $users as $user) {
-			$displayNames[$user] = $user;
-		}
-		return $displayNames;
-	}
-
-	/**
-	 * Check if a user list is available or not
-	 * @return boolean if users can be listed or not
-	 */
-	public function hasUserListings() {
-		return false;
-	}
 }
